ajustes de build
